+++ embc-app/ClientApp/src/app/shared/modals/session-expiring/session-expiring.component.ts
